Move dump_json into util_functions.py
